Installer: Remove references to the <DBPREFIX>settings table
Apparently, the settings table is no longer present in the structure SQL dump.
This broke the installer when verifying the tables present in the database, as the table name was still listed in the installer configuration file.
Note the TODO about the domainURL setting, the configureCaching() method, and the settings.php file itself in the _createSettingsFile() method.