Fix PHP notices about missing language entries
The EventListeners either did not load the language at all, or were using indices that did not exist.
Note that the GalleryEventListener included the Access ID array within the array of directories.
Also note that the new language entries probably need to be added to the backend and other languages, too.